chore: move config to blocks
