+++ blocks/Config/Config.js
@@ -1,10 +1,10 @@
 // @ts-check
-import { Block } from './Block.js';
-import { initialConfig } from './initialConfig.js';
-import { sharedConfigKey } from './sharedConfigKey.js';
+import { Block } from '../../abstract/Block.js';
+import { initialConfig } from '../../abstract/initialConfig.js';
+import { sharedConfigKey } from '../../abstract/sharedConfigKey.js';
 
 export class Config extends Block {
-  /** @type {Block['init$'] & import('./initialConfig.js').Config} */
+  /** @type {Block['init$'] & import('../../abstract/initialConfig.js').Config} */
   init$ = {
     ...this.init$,
     ...Object.fromEntries(Object.entries(initialConfig).map(([key, value]) => [sharedConfigKey(key), value])),
@@ -17,10 +17,10 @@ export class Config extends Block {
       Object.defineProperty(this, key, {
         /** @param {unknown} value */
         set: (value) => {
-          this.$[key] = value;
+          this.$[sharedConfigKey(key)] = value;
         },
         get: () => {
-          return this.$[key];
+          return this.$[sharedConfigKey(key)];
         },
       });
     });
