+++ blocks/Config/config.css
@@ -1,0 +1,3 @@
+lr-config {
+  display: none;
+}
